// ANGULAR/ANGULAR/src/app/Shared/Model/Master/monitoring-parameter-lookup-model.ts
export class MonitoringParameterlookupModel {
     Id: number;
     Name: string;
     NameHindi: string;
     LookupTypeCode: string;
     TypeCode: string;
}

export class MonitoringParameterLookupViewModel {
     lookuptypeName: string;
     lookuptypeNameinHindi: string;
     Name: string;
     NameHindi: string;
     TypeCode: string;
     Id: number;
}

export class MonitoringParameterLookupTypeModel{
     Id: number;
     LookupTypeCode: string;
     Name: string;
     NameHindi: string;
}